added services getCustomerAttrTypes  and  getGenParamsByHeader in CustomerService

+++ src/main/java/ge/softgen/loanexpert/model/CustomerAttrType.java
@@ -1,10 +1,8 @@
 package ge.softgen.loanexpert.model;
 
-import ge.softgen.loanexpert.spring.ApplicationConfig;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * Created by Bacho on 3/6/15.
@@ -14,18 +12,18 @@ import java.math.BigDecimal;
 public class CustomerAttrType implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	private BigDecimal id;
+	private Integer id;
 	private String descrip;
-	private BigDecimal attrtype;
+	private Integer attrtype;
 	private String attrlist;
 
 	@Id
 	@Column(name = "ID")
-	public BigDecimal getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(BigDecimal id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -41,11 +39,11 @@ public class CustomerAttrType implements Serializable {
 
 	@Basic
 	@Column(name = "ATTRTYPE")
-	public BigDecimal getAttrtype() {
+	public Integer getAttrtype() {
 		return attrtype;
 	}
 
-	public void setAttrtype(BigDecimal attrtype) {
+	public void setAttrtype(Integer attrtype) {
 		this.attrtype = attrtype;
 	}
